Итоги

+++ Homework 1/Дз1.3.py
@@ -1,4 +1,4 @@
-x = int(input("Введите х"))
-y = int(input("Введите у"))
-z = int(input("Введите z"))
+x = float(input("Введите х"))
+y = float(input("Введите у"))
+z = float(input("Введите z"))
 print((x ** 2 + y ** 2 + z ** 2) ** 0.5)
